Overlay: FFT correlation for the fast-vs-fastrx lag; dual-file test
A direct correlation of two 2.44 MHz pulse windows is minutes per
pulse. The dual path (paired fast pulse, lag) now has a test that
captures one event on both streams and a recording of it.

+++ test/fastrx_helpers.py
@@ -74,7 +74,7 @@ def record(mask: int, seq: int, *, snapshot=None, serial=42,
             assert block.shape == (SPP, 2)
             blocks += block.tobytes()
             continue
-        value = 100 * (p + 1) + seq
+        value = np.int64(100 * (p + 1) + seq).astype(np.int16)  # wraps
         samples = np.empty(2 * SPP, dtype=np.int16)
         samples[0::2] = value       # I
         samples[1::2] = -value      # Q
